Revert "1.16.5 duck helmet"
