+++ gradle.properties
@@ -4,11 +4,13 @@ org.gradle.jvmargs=-Xmx3G
 org.gradle.daemon=false
 
 # Mod Information
-mod_version=2.0.0-a
+mod_version=1.1.1
 
-# Dependencies
+# Minecraft Information
 minecraft_version=1.16.5
-curios_version=1.16.5-4.0.5.3
+
+#Forge Information (Remove on next version)
+forge_version=36.2.0
 
 #If you are forking this repository, put here your JAVA_HOME, dont forget to use double '\\'
 org.gradle.java.home=C:\\Program Files\\Eclipse Foundation\\jdk-8.0.302.8-hotspot
