+++ src/main/java/com/github/salvadormg15/rubber_duck/ForgeEventHandler.java
@@ -1,9 +1,6 @@
-package com.github.salvadormg15.rubber_duck.common.core;
+package com.github.salvadormg15.rubber_duck;
 
 import java.util.Random;
-
-import com.github.salvadormg15.rubber_duck.common.RubberDuck;
-import com.github.salvadormg15.rubber_duck.common.RubberDuckBlock;
 
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.item.ItemEntity;
