+++ src/main/java/com/github/salvadormg15/rubber_duck/RubberDuck.java
@@ -1,0 +1,35 @@
+package com.github.salvadormg15.rubber_duck;
+
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
+import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+// The value here should match an entry in the META-INF/mods.toml file
+@Mod("rubber_duck")
+public class RubberDuck
+{
+    // Directly reference a log4j logger.
+    public static final Logger LOGGER = LogManager.getLogger();
+    public static final String MODID = "rubber_duck";
+
+    public RubberDuck() {
+        IEventBus bus = FMLJavaModLoadingContext.get().getModEventBus();
+    	bus.addListener(this::setup);
+    	Registries.BLOCKS.register(bus);
+    	Registries.ITEMS.register(bus);
+    	Registries.SOUND_EVENTS.register(bus);
+
+        MinecraftForge.EVENT_BUS.register(this);
+        MinecraftForge.EVENT_BUS.register(ForgeEventHandler.class);
+    }
+
+    private void setup(final FMLCommonSetupEvent event)
+    {
+    }
+}
+
+
