+++ src/main/resources/META-INF/mods.toml
@@ -5,7 +5,7 @@ issueTrackerURL="https://github.com/salvadormg15/Rubber-Duck/issues"
 
 [[mods]]
     modId="rubber_duck"
-    version="2.0.0-a"
+    version="1.1.1"
     displayName="Rubber Duck"
     updateJSONURL="https://gist.githubusercontent.com/salvadormg15/f73e087096fbc1718d101c6d61389352/raw/"
     displayURL="https://www.curseforge.com/minecraft/mc-mods/rubber-duck"
@@ -23,11 +23,5 @@ issueTrackerURL="https://github.com/salvadormg15/Rubber-Duck/issues"
     modId="minecraft"
     mandatory=true
     versionRange="[1.16.5,1.17)"
-    ordering="NONE"
-    side="BOTH"
-[[dependencies.rubber_duck]]
-    modId="curios"
-    mandatory=true
-    versionRange="[1.16.5-4.0.5.3,)"
     ordering="NONE"
     side="BOTH"
