--- src/main/resources/data/curios/tags/items/head.json
@@ -1,4 +1,0 @@
-{
-	"replace": "false",
-	"values": ["rubber_duck:rubber_duck_item"]
-}
