+++ src/main/resources/data/rubber_duck/loot_tables/chests/rubber_duck.json
@@ -10,7 +10,7 @@
           "conditions": [
             {
               "condition": "minecraft:random_chance",
-              "chance": 0.2
+              "chance": 0.5
             }
           ]
         }
